+++ playground/islands/CHANGELOG.md
@@ -1,5 +1,12 @@
 # islands-spike
 
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [bde3a2d]
+  - @apex-stack/core@0.7.8
+
 ## 0.0.17
 
 ### Patch Changes